+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.18.0-beta.1 (Unreleased)
+## 12.18.0-beta.1 (2023-08-08)
 - Added support for service version 2023-05-03 and 2023-08-03.
 - Added RehydratePendingToCold value to ArchiveStatus enum.